+++ src/main/java/baseball/controller/Generator.java
@@ -1,6 +1,5 @@
-package baseball;
+package baseball.controller;
 
-import baseball.controller.Validation;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
